fix: improve test deletion UX with optimistic update
