+++ src/services/tests.service.js
@@ -26,6 +26,17 @@ class TestService {
     return tests;
   }
 
+
+
+  async deleteTest(id) {
+  const deleted = await this.testRepository.deleteTest(id);
+  if (!deleted) throw new AppError("Test not found", 404);
+  return deleted;
+}
+
+
+  
+
   async updateTest(id, updateData) {
     const test = await this.testRepository.updateTest(id, updateData);
     if (!test) throw AppError("Failed to update ", 500);
